feat(nestjs-jetstream): restructure registration options and mark legacy as deprecated
Mark the following legacy registration options as deprecated, to be removed in the next major release:
- streamName (use stream.name instead)
- durableName (use consumerOptions.name instead)
- deliverPolicy (use consumerOptions.deliver_policy instead)
- ackPolicy (use consumerOptions.ack_policy instead)
- ackWait (use consumerOptions.ack_wait instead)
- filterSubject (use consumerOptions.filter_subject instead)
- filterSubjects (use consumerOptions.filter_subjects instead)

Added @deprecated JSDoc annotations to the interface definitions and updated the README.md to document the deprecated options and provide migration guidance. Created a CHANGELOG.md file to track changes

BREAKING CHANGE: These options will be removed in the next major release. Users should migrate to the structured options (stream and consumerOptions) as soon as possible